+++ source/isaaclab_tasks/isaaclab_tasks/direct/constrained_albc/encoder/__init__.py
@@ -8,5 +8,11 @@
 from .actor_critic_constrained import ActorCriticConstrained
 from .actor_critic_encoder import ActorCriticEncoder
 from .actor_critic_encoder_constrained import ActorCriticEncoderConstrained
+from .actor_critic_frozen_encoder import ActorCriticFrozenEncoder
 
-__all__ = ["ActorCriticConstrained", "ActorCriticEncoder", "ActorCriticEncoderConstrained"]
+__all__ = [
+    "ActorCriticConstrained",
+    "ActorCriticEncoder",
+    "ActorCriticEncoderConstrained",
+    "ActorCriticFrozenEncoder",
+]